reorganized and renamed components
