+++ frontend/src/app/components/routing/render-on-anonymous.jsx
