+++ frontend/src/app/constants/roles.tsx
@@ -2,9 +2,9 @@
  * Roles enum, provide the available role names that a user can belong to.
  */
 export enum Roles {
-    SYSTEM_ADMINISTRATOR = 'System Administrator',
-    COS_ADMINISTRATOR = 'COS Administrator',
-    COS_OFFICER = 'COS Officer',
-  }
-  
-  export default Roles;
+  SYSTEM_ADMINISTRATOR = "System Administrator",
+  COS_ADMINISTRATOR = "COS Administrator",
+  COS_OFFICER = "COS Officer",
+}
+
+export default Roles;
